frontend/requirements_api.txt
flask[async]
flask-cors
google-api-python-client
google-cloud-bigquery
python-dotenv
